+++ sdk_nnbd/lib/_http/http_headers.dart
@@ -6,6 +6,8 @@ part of dart._http;
 
 class _HttpHeaders implements HttpHeaders {
   final Map<String, List<String>> _headers;
+  // The original header names keyed by the lowercase header names.
+  Map<String, String> _originalHeaderNames;
   final String protocolVersion;
 
   bool _mutable = true; // Are the headers currently mutable?
@@ -38,10 +40,10 @@ class _HttpHeaders implements HttpHeaders {
     }
   }
 
-  List<String>? operator [](String name) => _headers[name.toLowerCase()];
+  List<String>? operator [](String name) => _headers[_validateField(name)];
 
   String? value(String name) {
-    name = name.toLowerCase();
+    name = _validateField(name);
     List<String>? values = _headers[name];
     if (values == null) return null;
     assert(values.isNotEmpty);
@@ -51,13 +53,21 @@ class _HttpHeaders implements HttpHeaders {
     return values[0];
   }
 
-  void add(String name, value) {
+  void add(String name, value, {bool preserveHeaderCase: false}) {
     _checkMutable();
-    _addAll(_validateField(name), value);
+    String lowercaseName = _validateField(name);
+
+    // TODO(zichangguo): Consider storing all previously cased name for each
+    // entity instead of replacing name.
+    if (preserveHeaderCase && name != lowercaseName) {
+      (_originalHeaderNames ??= {})[lowercaseName] = name;
+    } else {
+      _originalHeaderNames?.remove(lowercaseName);
+    }
+    _addAll(lowercaseName, value);
   }
 
   void _addAll(String name, value) {
-    assert(name == _validateField(name));
     if (value is Iterable) {
       for (var v in value) {
         _add(name, _validateValue(v));
@@ -67,28 +77,35 @@ class _HttpHeaders implements HttpHeaders {
     }
   }
 
-  void set(String name, Object value) {
+  void set(String name, Object value, {bool preserveHeaderCase: false}) {
     _checkMutable();
-    name = _validateField(name);
-    _headers.remove(name);
-    if (name == HttpHeaders.transferEncodingHeader) {
+    String lowercaseName = _validateField(name);
+    _headers.remove(lowercaseName);
+    _originalHeaderNames?.remove(lowercaseName);
+    if (lowercaseName == HttpHeaders.transferEncodingHeader) {
       _chunkedTransferEncoding = false;
     }
-    _addAll(name, value);
+    if (preserveHeaderCase && name != lowercaseName) {
+      (_originalHeaderNames ??= {})[lowercaseName] = name;
+    }
+    _addAll(lowercaseName, value);
   }
 
   void remove(String name, Object value) {
     _checkMutable();
     name = _validateField(name);
     value = _validateValue(value);
-    if (name == HttpHeaders.transferEncodingHeader && value == "chunked") {
-      _chunkedTransferEncoding = false;
-      return;
-    }
     List<String>? values = _headers[name];
     if (values != null) {
       values.remove(_valueToString(value));
-      if (values.length == 0) _headers.remove(name);
+      if (values.length == 0) {
+        _headers.remove(name);
+        _originalHeaderNames?.remove(name);
+      }
+    }
+    if (name == HttpHeaders.transferEncodingHeader && value == "chunked") {
+      _chunkedTransferEncoding = false;
+      return;
     }
   }
 
@@ -96,10 +113,14 @@ class _HttpHeaders implements HttpHeaders {
     _checkMutable();
     name = _validateField(name);
     _headers.remove(name);
+    _originalHeaderNames?.remove(name);
   }
 
-  void forEach(void f(String name, List<String> values)) {
-    _headers.forEach(f);
+  void forEach(void action(String name, List<String> values)) {
+    _headers.forEach((String name, List<String> values) {
+      String originalName = _originalHeaderName(name);
+      action(originalName, values);
+    });
   }
 
   void noFolding(String name) {
@@ -511,14 +532,15 @@ class _HttpHeaders implements HttpHeaders {
   String toString() {
     StringBuffer sb = new StringBuffer();
     _headers.forEach((String name, List<String> values) {
-      sb..write(name)..write(": ");
+      String originalName = _originalHeaderName(name);
+      sb..write(originalName)..write(": ");
       bool fold = _foldHeader(name);
       for (int i = 0; i < values.length; i++) {
         if (i > 0) {
           if (fold) {
             sb.write(", ");
           } else {
-            sb..write("\n")..write(name)..write(": ");
+            sb..write("\n")..write(originalName)..write(": ");
           }
         }
         sb.write(values[i]);
@@ -604,7 +626,7 @@ class _HttpHeaders implements HttpHeaders {
     for (var i = 0; i < field.length; i++) {
       if (!_HttpParser._isTokenChar(field.codeUnitAt(i))) {
         throw new FormatException(
-            "Invalid HTTP header field name: ${json.encode(field)}");
+            "Invalid HTTP header field name: ${json.encode(field)}", field, i);
       }
     }
     return field.toLowerCase();
@@ -615,10 +637,14 @@ class _HttpHeaders implements HttpHeaders {
     for (var i = 0; i < value.length; i++) {
       if (!_HttpParser._isValueChar(value.codeUnitAt(i))) {
         throw new FormatException(
-            "Invalid HTTP header field value: ${json.encode(value)}");
+            "Invalid HTTP header field value: ${json.encode(value)}", value, i);
       }
     }
     return value;
+  }
+
+  String _originalHeaderName(String name) {
+    return _originalHeaderNames?.[name] ?? name;
   }
 }